Move inferred tests to their own suite
